Make FatFile::read() asynchronous, plus deps.
Deps:

- Add a Future.h, which started with Boost's future (since it had the
option to do .then(), while C++14 did not).  Made tons of changes
though: removed all threading, removed all async, ripped out lots of the
shared state, removed the promise.  Added an option for a callback to
determine if a future is done, and inverted most of the triggering of
continuations to be based off is_done() polling instead of automatic
invoking of long running operations.  Documented which APIs are needed,
and made sure they work.

- Adapt FatLib/BaseBlockDriver to return futures, for the base API.

- Adapt all the low level SPI drivers to return a future instead of a
direct result.  The drivers that don't do DMA do their old code and then
return a ready future; the ones that do DMA set up a callback to check
on the DMA-done flag (that the IRQ handler sets).

- Adapt the SDIO drivers to do the same DMA.

- Adapt SdCard drivers to use the futures and operate asynchronously,
for reads only.

- Adapt FatLib/FatVolume to make the caching return a future instead of
the pointer directly.  This only works because there's only ever one
block in the cache.

- Adapt FatLib/FatFile to use the futures from both the block driver and
the FatVolume code.  Most of the time we just call .get() and make it
synchronous.  But read() requires significant restructuring.
  - Because read() issues many different types of read operations to the
    cache and the underlying block driver, it needs to preprocess the
    read setup; first it builds a list of the reads that will need to be
    issued (including where in the buffer they get written to), then it
    runs through that list and sets up a sequence of .then continuations
    to actually issue the reads and process the results.
  - Then all the way down at the end we modify the current position.

- Adapt FatLib/StdioStream and FatLib/fstream to the asynchronous setup:
make them process the reads synchronously because the API that fstream
and stdio provide can't be async.  Using ::read() directly is enough for
what I need to do.

This builds.  I don't yet know if it works.  :)