Use IRQ builtins, work around #define F() in hdrs.
Use HAL builtins to access PRIMASK and to enable/disable interrupts in
the lock, instead of manually generating the inline assembly.

When compiling Future.h in a full particle project, the Particle.h
header (which is hardcoded #include<>d at the start) includes a
spark_wiring_string.h header, which has a "#define F(x) (x)" line.  This
completely breaks the old Future code, which tried to invoke
std::result_of<F(future)>::type (since it had an F template parameter
that was a function taking this future as an argument).  Rename F to
Func.  Sigh.