Fix compilation, fix API comment
Costants need to be outside any function to be usable from a lambda
without explicitly capturing them, it seems, so move them out.

Using memset on a non-POD class is undefined, and the constructor zero
initializes things for us already, so drop the memset.

Several references to "block" did not get changed to "d.block" in the
previous changes; fix them up and delete the block local variable.

Explicitly include the NRF header from Future, since the new code to
disable and enable interrupts requires that header.

Fix the API comment to return a new future (as the function does)
instead of trying to modify the existing one's return value.