Make futures actually work.
Turns out I really do need to use most of the structure from the boost
future implementation that I started with, including the three levels of
inheritance in the shared-state tree (and with template types on the
second and third level).  One level handles continuations only; the next
handles the done bit and the result; the third handles the prereqs for a
continuation.

While I'm in here, clean up the APIs a bit; no point in having duplicate
public functions.  Make the Particle code call the correct new function.

With this, the microcontroller boots, reads the FAT metadata off the SD
card, and successfully initializes the filesystem instance.